[Refactor] Admin -> Manager 1차 변경
