+++ src/docs/asciidoc/manager/manager-api.adoc
@@ -1,0 +1,39 @@
+= Admin REST API Docs
+:doctype: book
+:icons: font
+:source-highlighter: highlightjs
+:toc: left
+:toclevels: 2
+
+=== link:../index.html[시작 페이지로 돌아가기]
+
+[[Admin-Register]]
+== 관리자 회원가입
+
+관리자 회원가입 API 입니다.
+
+=== HttpRequest
+
+include::{snippets}/manager-register/http-request.adoc[]
+include::{snippets}/manager-register/request-fields.adoc[]
+
+=== Http Response
+include::{snippets}/manager-register/http-response.adoc[]
+include::{snippets}/manager-register/response-fields.adoc[]
+
+[[Admin-Login]]
+== 관리자 로그인
+
+관리자 로그인 API 입니다.
+
+* 발급된 액세스 토큰은 헤더에 저장됩니다.
+
+=== HttpRequest
+
+include::{snippets}/manager-login/http-request.adoc[]
+include::{snippets}/manager-login/request-fields.adoc[]
+
+=== Http Response
+include::{snippets}/manager-login/http-response.adoc[]
+include::{snippets}/manager-login/response-fields.adoc[]
+include::{snippets}/manager-login/response-headers.adoc[]
