+++ src/main/java/com/final_10aeat/common/enumclass/MemberRole.java
@@ -1,5 +1,5 @@
 package com.final_10aeat.common.enumclass;
 
 public enum MemberRole {
-    ADMIN, OWNER, TENANT
+    MANAGER, OWNER, TENANT
 }
