--- src/main/java/com/final_10aeat/domain/admin/entity/Admin.java
@@ -1,66 +1,0 @@
-package com.final_10aeat.domain.admin.entity;
-
-import com.final_10aeat.common.enumclass.MemberRole;
-import com.final_10aeat.domain.office.entity.Office;
-import com.final_10aeat.global.entity.SoftDeletableBaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@Table(name = "admin")
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Admin extends SoftDeletableBaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true)
-    private String email;
-
-    @Column
-    private String password;
-
-    @Column
-    private String name;
-
-    @Column(name = "phonenumber")
-    private String phoneNumber;
-
-    @Column(name = "lunch_start")
-    private LocalDateTime lunchBreakStart;
-
-    @Column(name = "lunch_end")
-    private LocalDateTime lunchBreakEnd;
-
-    @Column(name = "admin_office")
-    private String adminOffice;
-
-    private String affiliation;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private MemberRole role = MemberRole.ADMIN;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "office_id")
-    private Office office;
-}
