--- src/main/java/com/final_10aeat/domain/admin/exception/InvalidVerificationCodeException.java
@@ -1,14 +1,0 @@
-package com.final_10aeat.domain.admin.exception;
-
-import com.final_10aeat.global.exception.ApplicationException;
-import com.final_10aeat.global.exception.ErrorCode;
-
-public class InvalidVerificationCodeException extends ApplicationException {
-
-    private static final ErrorCode ERROR_CODE = ErrorCode.INVALID_VERIFICATION_CODE;
-
-    public InvalidVerificationCodeException() {
-
-        super(ERROR_CODE);
-    }
-}
