--- src/main/java/com/final_10aeat/domain/admin/repository/AdminRepository.java
@@ -1,10 +1,0 @@
-package com.final_10aeat.domain.admin.repository;
-
-import com.final_10aeat.domain.admin.entity.Admin;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AdminRepository extends JpaRepository<Admin, Long> {
-
-    Optional<Admin> findByEmail(String email);
-}
