+++ src/main/java/com/final_10aeat/domain/manager/controller/EmailController.java
@@ -1,0 +1,45 @@
+package com.final_10aeat.domain.manager.controller;
+
+import com.final_10aeat.domain.manager.dto.request.EmailRequestDto;
+import com.final_10aeat.domain.manager.dto.request.EmailVerificationRequestDto;
+import com.final_10aeat.domain.manager.dto.response.EmailVerificationResponseDto;
+import com.final_10aeat.domain.manager.service.ManagerService;
+import com.final_10aeat.domain.manager.service.EmailUseCase;
+import com.final_10aeat.global.security.principal.ManagerPrincipal;
+import com.final_10aeat.global.util.ResponseDTO;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/members/email")
+public class EmailController {
+
+    private final EmailUseCase emailUseCase;
+    private final ManagerService managerService;
+
+    @PostMapping
+    // @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<ResponseDTO<Void>> mailConfirm(
+        @RequestBody @Valid EmailRequestDto emailRequest,
+        @AuthenticationPrincipal ManagerPrincipal managerPrincipal) throws Exception {
+        Long officeId = managerPrincipal.getManager().getOffice().getId();
+        emailUseCase.sendVerificationEmail(emailRequest.email(), emailRequest.role(),
+            emailRequest.dong(), emailRequest.ho(), officeId);
+        return ResponseEntity.ok(ResponseDTO.ok());
+    }
+
+    @PostMapping("/verification")
+    public ResponseEntity<ResponseDTO<EmailVerificationResponseDto>> verifyEmail(
+        @RequestBody @Valid EmailVerificationRequestDto verificationRequest) {
+        EmailVerificationResponseDto responseDto = emailUseCase.verifyEmailCode(
+            verificationRequest.email(), verificationRequest.code());
+        return ResponseEntity.ok(ResponseDTO.okWithData(responseDto));
+    }
+}
