+++ src/main/java/com/final_10aeat/domain/manager/dto/response/EmailVerificationResponseDto.java
@@ -1,0 +1,18 @@
+package com.final_10aeat.domain.manager.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class EmailVerificationResponseDto {
+
+    private String email;
+    private String role;
+    private String dong;
+    private String ho;
+}
