+++ src/main/java/com/final_10aeat/global/config/SecurityConfig.java
@@ -35,7 +35,7 @@ public class SecurityConfig {
             .authorizeHttpRequests(
                 auth -> auth
                     .requestMatchers(HttpMethod.POST, "/members", "/members/login",
-                        "/admin","/admin/login","/members/email/verification").permitAll()
+                        "/manager","/manager/login","/members/email/verification").permitAll()
                     .requestMatchers(HttpMethod.GET, "/health/**").permitAll()
                     .requestMatchers("/docs/**").permitAll()
                     .anyRequest().authenticated()
