--- src/main/java/com/final_10aeat/global/security/principal/AdminPrincipal.java
@@ -1,50 +1,0 @@
-package com.final_10aeat.global.security.principal;
-
-import com.final_10aeat.domain.admin.entity.Admin;
-import java.util.Collection;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-@RequiredArgsConstructor
-public class AdminPrincipal implements UserDetails {
-
-    @Getter
-    private final Admin admin;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public String getPassword() {
-        return admin.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return admin.getEmail();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return false;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return false;
-    }
-}
